refactor: re-export index.ts
